% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_1_source.R
\name{gradient.descent.BB.lsq_R}
\alias{gradient.descent.BB.lsq_R}
\title{use BB method to determine adaptive step sizes
method is 'lasso' or 'ridge'}
\usage{
gradient.descent.BB.lsq_R(
  y,
  A,
  x0,
  lambda,
  tol = 1e-04,
  max.iter = 10000,
  printing = FALSE
)
}
\description{
use BB method to determine adaptive step sizes
method is 'lasso' or 'ridge'
}
