% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_1_source.R
\name{stochastic.gradient.descent.lsq_R}
\alias{stochastic.gradient.descent.lsq_R}
\title{method is 'lasso' or 'ridge'
batch is batch size, ranging from 1 to n (original gradient descent)}
\usage{
stochastic.gradient.descent.lsq_R(
  y,
  A,
  x0,
  lambda,
  batch,
  initial.step.size = 1,
  tol = 1e-06,
  max.iter = 10000,
  printing = FALSE
)
}
\description{
method is 'lasso' or 'ridge'
batch is batch size, ranging from 1 to n (original gradient descent)
}
